requirements.txt
certifi==2019.11.28
cffi==1.13.2
cryptography==2.8
future==0.18.2
phonenumbers==8.11.1
pycountry==19.8.18
pycparser==2.19
python-telegram-bot==12.2.0
six==1.13.0
tornado==6.0.3
